+++ CHANGELOG.md
@@ -59,7 +59,9 @@
 - Changes to xSQLServerDatabaseOwner
   - Fixed code style, added updated parameter descriptions to schema.mof and README.md.
 - Changes to xSQLServerRole
-  - BREAKING CHANGE: The mandatory parameters now include SQLServer and ServerRole. The new parameter are Members, MembersToInclude and MembersToExclude.
+  - BREAKING CHANGE: The mandatory parameters now include SQLServer and ServerRoleName. The new parameter are Members, MembersToInclude and MembersToExclude.
+  - The ServerRole parameter was before an array of server roles, now this parameter is renamed by ServerRoleName and contains only one server role to set.
+  - The new parameter Members, MembersToInclude and MembersToExclude contains one or multiple logins that can be adding or dropping in server role members.
 
 ## 5.0.0.0